move auth attributes into a folder
