+++ backend/LexBoxApi/Controllers/MigrationController.cs
@@ -1,4 +1,5 @@
 using LexBoxApi.Auth;
+using LexBoxApi.Auth.Attributes;
 using LexBoxApi.Services;
 using LexCore.Entities;
 using LexData;
